+++ src/main/java/filter/ParseException.java
@@ -1,8 +1,12 @@
 package filter;
 
+/**
+ *  Thrown to indicate failure to parse a filter expression due to 
+ *  unsupported syntax
+ */
 @SuppressWarnings("serial")
-public class ParseException extends RuntimeException {
+public class ParseException extends FilterException {
       public ParseException(String message) {
-         super(message);
+         super("Error in filter: " + message);
       }
  }