+++ src/test/java/guitests/UITest.java
@@ -1,16 +1,19 @@
 package guitests;
 
-import backend.interfaces.RepoStore;
-import com.google.common.util.concurrent.SettableFuture;
-import javafx.geometry.Point2D;
-import javafx.scene.Node;
-import javafx.scene.Parent;
-import javafx.scene.control.ComboBoxBase;
-import javafx.scene.input.KeyCode;
-import javafx.scene.input.KeyCodeCombination;
-import javafx.scene.input.KeyCombination;
-import javafx.scene.input.MouseButton;
-import javafx.stage.Stage;
+import static com.google.common.io.Files.getFileExtension;
+
+import java.awt.Robot;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.*;
+import java.util.concurrent.Callable;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
+
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -21,24 +24,30 @@ import org.loadui.testfx.GuiTest;
 import org.loadui.testfx.exceptions.NoNodesFoundException;
 import org.loadui.testfx.exceptions.NoNodesVisibleException;
 import org.loadui.testfx.utils.FXTestUtils;
+
+import com.google.common.util.concurrent.SettableFuture;
+
+import backend.interfaces.RepoStore;
+import javafx.application.Platform;
+import javafx.event.ActionEvent;
+import javafx.geometry.Point2D;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.control.ComboBoxBase;
+import javafx.scene.control.Menu;
+import javafx.scene.control.MenuItem;
+import javafx.scene.control.TextField;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyCodeCombination;
+import javafx.scene.input.KeyCombination;
+import javafx.scene.input.MouseButton;
+import javafx.stage.Stage;
 import prefs.Preferences;
+import ui.MenuControl;
+import ui.TestController;
 import ui.UI;
 import util.PlatformEx;
 import util.PlatformSpecific;
-
-import java.awt.*;
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.*;
-import java.util.List;
-import java.util.concurrent.Callable;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.function.Predicate;
-
-import static com.google.common.io.Files.getFileExtension;
 
 public class UITest extends GuiTest {
 
@@ -92,7 +101,7 @@ public class UITest extends GuiTest {
         return Collections.unmodifiableMap(specialChars);
     }
 
-    public void setupMethod() {
+    public void beforeStageStarts() {
         // method to be overridden if anything needs to be done (e.g. to the json) before the stage starts
     }
 
@@ -100,12 +109,12 @@ public class UITest extends GuiTest {
         try {
             if (Files.exists(Paths.get(RepoStore.TEST_DIRECTORY))) {
                 Files.walk(Paths.get(RepoStore.TEST_DIRECTORY), 1)
-                        .filter(Files::isRegularFile)
-                        .filter(p ->
-                                getFileExtension(String.valueOf(p.getFileName())).equalsIgnoreCase("json") ||
-                                getFileExtension(String.valueOf(p.getFileName())).equalsIgnoreCase("json-err")
-                        )
-                        .forEach(p -> new File(p.toAbsolutePath().toString()).delete());
+                    .filter(Files::isRegularFile)
+                    .filter(p ->
+                        getFileExtension(String.valueOf(p.getFileName())).equalsIgnoreCase("json") ||
+                            getFileExtension(String.valueOf(p.getFileName())).equalsIgnoreCase("json-err")
+                    )
+                    .forEach(p -> new File(p.toAbsolutePath().toString()).delete());
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -121,7 +130,7 @@ public class UITest extends GuiTest {
             assert testConfig.delete();
         }
         clearTestFolder();
-        setupMethod();
+        beforeStageStarts();
 
         if (stage == null) {
             launchApp();
@@ -246,38 +255,25 @@ public class UITest extends GuiTest {
     }
 
     public void waitUntilNodeAppears(String selector) {
-        while (!existsQuiet(selector)) {
-            PlatformEx.waitOnFxThread();
-            sleep(100);
-        }
+        awaitCondition(() -> existsQuiet(selector));
     }
 
     public void waitUntilNodeDisappears(String selector) {
-        while (existsQuiet(selector)) {
-            PlatformEx.waitOnFxThread();
-            sleep(100);
-        }
+        awaitCondition(() -> !existsQuiet(selector));
     }
 
     public void waitUntilNodeAppears(Matcher<Object> matcher) {
-        while (!findQuiet(matcher).isPresent()) { // no `exists` for matcher, so using find
-            PlatformEx.waitOnFxThread();
-            sleep(100);
-        }
+        // We use find because there's no `exists` for matchers
+        awaitCondition(() -> findQuiet(matcher).isPresent());
     }
 
     public void waitUntilNodeDisappears(Matcher<Object> matcher) {
-        while (findQuiet(matcher).isPresent()) { // no `exists` for matcher, so using find
-            PlatformEx.waitOnFxThread();
-            sleep(100);
-        }
+        // We use find because there's no `exists` for matchers
+        awaitCondition(() -> !findQuiet(matcher).isPresent());
     }
 
     public <T extends Node> void waitUntil(String selector, Predicate<T> condition) {
-        while (!condition.test(find(selector))) {
-            PlatformEx.waitOnFxThread();
-            sleep(100);
-        }
+        awaitCondition(() -> condition.test(find(selector)));
     }
 
     public <T extends Node> T findOrWaitFor(String selector) {
@@ -325,19 +321,19 @@ public class UITest extends GuiTest {
 
     /**
      * Allows test threads to busy-wait on some condition.
-     *
+     * <p>
      * Taken from org.loadui.testfx.utils, but modified to synchronise with
      * the JavaFX Application Thread, with a lower frequency.
-     *
+     * <p>
      * The additional synchronisation prevents bugs where
-     *
+     * <p>
      * awaitCondition(a);
      * awaitCondition(b);
-     *
+     * <p>
      * sometimes may not be equivalent to
-     *
+     * <p>
      * awaitCondition(a && b);
-     *
+     * <p>
      * The lower frequency is a bit more efficient, since a frequency of 10 ms
      * just isn't necessary for GUI interactions, and we're bottlenecked by the FX
      * thread anyway.
@@ -362,6 +358,36 @@ public class UITest extends GuiTest {
     }
 
     /**
+     * Performs logout from File -> Logout on HubTurbo's pView.
+     */
+    public void logout(){
+        clickMenu("File", "Logout");
+    }
+
+    /**
+     * Performs UI login on the login dialog box.
+     * @param owner The owner of the repo.
+     * @param repoName The repository name
+     * @param username The Github username
+     * @param password The Github password
+     */
+    public void login(String owner, String repoName, String username, String password){
+        selectAll();
+        type(owner).push(KeyCode.TAB);
+        type(repoName).push(KeyCode.TAB);
+        type(username).push(KeyCode.TAB);
+        type(password);
+        click("Sign in");
+    }
+
+    /**
+     * Waits for the result of a function, then asserts that it is equal to some value.
+     */
+    public <T> void waitAndAssertEquals(T expected, Supplier<T> actual) {
+        awaitCondition(() -> expected.equals(actual.get()));
+    }
+
+    /**
      * Automate menu traversal by clicking them in order of input parameter
      *
      * @param menuNames array of strings of menu item names in sequence of traversal
@@ -372,17 +398,71 @@ public class UITest extends GuiTest {
         }
     }
 
+    /**
+     * Sets a text field with given text. Does not simulate clicking and typing 
+     * in the text field.
+     * @param fieldId
+     * @param text
+     */
+    public void setTextField(String fieldId, String text) {
+        waitUntilNodeAppears(fieldId);
+        ((TextField) find(fieldId)).setText(text);
+    }
+
     public <T> void waitForValue(ComboBoxBase<T> comboBoxBase) {
         waitUntil(comboBoxBase, c -> c.getValue() != null);
     }
-    
+
+    /**
+     * Traverses HubTurbo's menu, looking for a chain of nodes with the
+     * given names and triggering their associated action.
+     *
+     * This is a more reliable method of triggering menu items than
+     * {@link #clickMenu}, especially when dealing with nested menu items.
+     * It is a drop-in replacement in most cases.
+     *
+     * Caveats: ensure that adequate synchronisation is used after this method
+     * if it is called from a thread other than the UI thread.
+     *
+     * @param names the chain of menu nodes to visit
+     */
+    public void traverseMenu(String... names) {
+        assert names.length > 0 : "traverseMenu called with no arguments";
+
+        Platform.runLater(() -> {
+            MenuControl root = TestController.getUI().getMenuControl();
+            MenuItem current = root.getMenus().stream()
+                .filter(m -> m.getText().equals(names[0]))
+                .findFirst()
+                .orElseThrow(() ->
+                    new IllegalArgumentException(
+                        String.format("%s is not a valid menu item", names[0])));
+
+            for (int i = 1; i < names.length; i++) {
+                final int j = i;
+                if (!(current instanceof Menu)) {
+                    throw new IllegalArgumentException(
+                        String.format("Menu %s is not as nested as arguments require", names[0]));
+                }
+                current = ((Menu) current).getItems().stream()
+                    .filter(m -> m.getText().equals(names[j]))
+                    .findFirst()
+                    .orElseThrow(() ->
+                        new IllegalArgumentException(
+                            String.format("%s is not a valid menu item", names[j])));
+            }
+
+            current.getOnAction().handle(new ActionEvent());
+        });
+    }
+
     @Override
     public GuiTest type(String text) {
         for (int i = 0; i < text.length(); i++) {
-            if (specialCharsMap.containsKey(text.charAt(i))){
+            if (specialCharsMap.containsKey(text.charAt(i))) {
                 press(KeyCode.SHIFT).press(specialCharsMap.get(text.charAt(i)))
-                .release(specialCharsMap.get(text.charAt(i))).release(KeyCode.SHIFT);
-             
+                    .release(specialCharsMap.get(text.charAt(i))).release(KeyCode.SHIFT);
+
             } else {
                 type(text.charAt(i));
             }