+++ src/test/java/tests/TurboMilestoneTest.java
@@ -5,6 +5,8 @@ import org.eclipse.egit.github.core.Milestone;
 import org.junit.Test;
 
 import java.time.LocalDate;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 
 import static org.junit.Assert.assertEquals;
@@ -29,6 +31,32 @@ public class TurboMilestoneTest {
         assertEquals(0, turboMilestone.getOpenIssues());
         turboMilestone.setClosedIssues(0);
         assertEquals(0, turboMilestone.getClosedIssues());
+    }
+
+    @Test
+    public void sortByDueDate_sameMilestoneDueDate_stableSorting() {
+        // test for turbo milestone due date sorting behaviour
+        TurboMilestone milestone1 = new TurboMilestone("1", 1, "milestone1");
+        milestone1.setDueDate(Optional.of(LocalDate.now().minusDays(1)));
+        TurboMilestone milestone2 = new TurboMilestone("2", 2, "milestone2");
+        milestone2.setDueDate(Optional.of(LocalDate.now().minusDays(2)));
+        TurboMilestone milestone3 = new TurboMilestone("3", 3, "milestone3");
+        milestone3.setDueDate(Optional.of(LocalDate.now().minusDays(1)));
+        TurboMilestone milestone4 = new TurboMilestone("3", 4, "milestone4");
+        TurboMilestone milestone5 = new TurboMilestone("3", 5, "milestone5");
+
+        // the sort should be stable
+        List<TurboMilestone> milestones = Arrays.asList(milestone1,
+                milestone2,
+                milestone3,
+                milestone4,
+                milestone5);
+        List<TurboMilestone> sortedMilestones = TurboMilestone.sortByDueDate(milestones);
+        assertEquals("milestone2", sortedMilestones.get(0).getTitle());
+        assertEquals("milestone1", sortedMilestones.get(1).getTitle());
+        assertEquals("milestone3", sortedMilestones.get(2).getTitle());
+        assertEquals("milestone4", sortedMilestones.get(3).getTitle());
+        assertEquals("milestone5", sortedMilestones.get(4).getTitle());
     }
 
 }